Comments passing, reorganization
