+++ api/src/index.ts
@@ -2,3 +2,4 @@ export * from './Upwell'
 export * from './UpwellMetadata'
 export * from './Layer'
 export * from './RTC'
+export * from './Comments'
