+++ api/test/marks.ts
@@ -2,7 +2,7 @@ import { createAuthorId, CommentState, Upwell } from '../src/index'
 import { it } from 'mocha';
 import { assert } from 'chai';
 
-describe.only('upwell', () => {
+describe('upwell', () => {
   let author = {id: createAuthorId(), name: 'susan'}
   let d = Upwell.create({ author })
   it('works', async () => {
